+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 4.5.0 - Unreleased
 ### Added
 - Added the ability to send compressed responses to browsers that accept supported encodings.
-- Added the ability to saved compress cached values in the Yii Cache Storage to help reduce the amount of memory required.
+- Added the ability to save compressed cached values in the Yii Cache Storage to help reduce the memory required.
 - Added tips that display whether `Gzip` and `Brotli` are enabled on the web server in the Blitz File Storage settings.
 
 ### Fixed