+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Added
 - Added the ability to send compressed responses to browsers that accept supported encodings.
 - Added the ability to save compressed cached values in the Yii Cache Storage to help reduce the memory required.
-- Added tips that display whether `Gzip` and `Brotli` are enabled on the web server in the Blitz File Storage settings.
+- Added tips that display whether `gzip` and `Brotli` are enabled on the web server in the Blitz File Storage settings.
 
 ### Fixed
 - Fixed the `generatePageBasedOnQueryString` check and ensured that cached includes are always allowed.